printf done

+++ ft_printf/f_types/func_hex_ptr.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   func_hex_ptr.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jimlee <jimlee@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jimlee <jimlee@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/22 17:38:14 by jimlee            #+#    #+#             */
-/*   Updated: 2022/11/24 15:26:20 by jimlee           ###   ########.fr       */
+/*   Updated: 2022/11/29 17:44:05 by jimlee           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,11 +38,11 @@ int	print_hex_upper(t_format *format, va_list ap)
 
 int	print_ptr(t_format *format, va_list ap)
 {
-	unsigned long long	val;
-	int					put_size;
+	uintptr_t	val;
+	int			put_size;
 
 	(void)format;
-	val = va_arg(ap, unsigned long long);
+	val = va_arg(ap, uintptr_t);
 	put_size = 2 + get_len_num_u(val, 16);
 	write(STDOUT_FILENO, "0x", 2);
 	ft_putnbr_base_u(val, "0123456789abcdef");
